requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
aiosmtplib==5.0.0
aiosqlite==0.22.1
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.1.5
apache-airflow-core==3.1.5
apache-airflow-providers-common-compat==1.11.0
apache-airflow-providers-common-io==1.7.0
apache-airflow-providers-common-sql==1.30.2
apache-airflow-providers-smtp==2.4.1
apache-airflow-providers-standard==1.10.2
apache-airflow-task-sdk==1.1.5
appnope==0.1.3
argcomplete==3.6.3
asgiref==3.11.0
asttokens==2.4.1
attrs==25.4.0
Automat==22.10.0
babel==2.17.0
cachetools==6.2.4
cadwyn==5.4.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.3.1
colorlog==6.10.1
comm==0.2.0
constantly==23.10.4
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==42.0.7
cssselect==1.2.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
executing==2.0.1
fastapi==0.117.1
fastapi-cli==0.0.20
filelock==3.14.0
frozenlist==1.7.0
fsspec==2025.12.0
googleapis-common-protos==1.72.0
gotrue==2.12.0
greenback==1.3.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.7.1
incremental==22.10.0
iniconfig==2.1.0
ipykernel==6.27.1
ipython==8.18.1
itemadapter==0.9.0
itemloaders==1.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.0
jupyter_core==5.5.1
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.2.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.6
mdurl==0.1.2
methodtools==0.4.7
mmh3==5.2.0
more-itertools==10.8.0
msgspec==0.20.0
multidict==6.4.4
natsort==8.4.0
nest-asyncio==1.5.8
nltk==3.9.2
openai==0.28.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
outcome==1.3.0.post0
packaging==25.0
parsel==1.9.1
parso==0.8.3
pathspec==1.0.1
pendulum==3.1.0
pexpect==4.9.0
platformdirs==4.1.0
pluggy==1.6.0
postgrest==2.27.1
prompt-toolkit==3.0.43
propcache==0.3.2
Protego==0.3.1
protobuf==6.33.2
psutil==7.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
PyDispatcher==2.0.7
Pygments==2.17.2
pygtrie==2.5.0
pyiceberg==0.10.0
PyJWT==2.10.1
PyMuPDF==1.26.1
pyOpenSSL==24.1.0
pyparsing==3.3.1
pyroaring==1.0.3
pytest==8.4.0
pytest-mock==3.14.1
python-daemon==3.1.2
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.21
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==25.1.2
queuelib==1.7.0
realtime==2.27.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==2.0.0
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.17.1
rpds-py==0.30.0
Scrapy==2.11.2
service-identity==24.1.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.48.0
storage3==2.27.1
StrEnum==0.4.15
strictyaml==1.7.3
structlog==25.5.0
supabase==2.27.1
supabase-py==2.27.1
supabase-auth==2.27.1
supabase-functions==2.27.1
supafunc==0.9.4
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
tldextract==5.1.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.0
Twisted==24.3.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.2.1
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
w3lib==2.1.2
watchfiles==1.1.1
wcwidth==0.2.12
websockets==14.2
wirerope==1.0.0
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
zope.interface==6.4
